Add files via upload
The python program contains a class "filescraper" that has multiple functions that maybe applied to an instance of the class filescraper. The program will take an input path of the form "User\example\example\..." to create the filescraper object which now has an associated folder. We may then run functions associated with the filescraper object to access any files within the  folder.